requirements.txt
alembic==1.8.0
attrs==21.4.0
blinker==1.4
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.4
commonmark==0.9.1
coverage==6.4.3
cryptography==37.0.2
dataclass-type-validator==0.1.2
defusedxml==0.7.1
dnspython==2.2.1
dominate==2.6.0
email-validator==1.2.1
Flask==2.1.2
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.7.0
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
httpie==3.2.1
idna==3.3
importlib-metadata==4.11.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
Mako==1.2.0
MarkupSafe==2.1.1
multidict==6.0.2
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pydantic==1.9.1
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
PySocks==1.7.1
pyTelegramBotAPI==4.6.0
pytest==7.1.2
pytest-cov==3.0.0
python-dotenv==0.20.0
requests==2.28.1
requests-toolbelt==0.9.1
rich==12.4.4
SQLAlchemy==1.4.37
sqlparse==0.3.0
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.9
visitor==0.1.3
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
